requirements.txt
streamlit
pandas
openpyxl
PyPDF2
google-generativeai
python-dotenv
xlsxwriter